backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Redis & Celery
redis==5.2.1
celery==5.4.0

# MQTT
gmqtt==0.6.16

# Utils
httpx==0.28.1
python-dateutil==2.9.0
email-validator==2.2.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
factory-boy==3.3.1
aiosqlite==0.20.0
